+++ cli/cmd/configure_bridge.go
@@ -1,18 +1,7 @@
 package cmd
 
 import (
-	"bytes"
-	"crypto/tls"
-	"encoding/json"
-	"errors"
 	"fmt"
-	"io/ioutil"
-	"net/http"
-	"net/url"
-	"path"
-	"strings"
-
-	"github.com/keptn/keptn/cli/pkg/credentialmanager"
 	"github.com/spf13/cobra"
 )
 
@@ -35,145 +24,24 @@ var bridgeCmd = &cobra.Command{
 	Long:         `Configures the credentials for the Keptn Bridge.`,
 	Example:      `keptn configure bridge --user=<user> --password=<password>`,
 	SilenceUsage: true,
-	PreRunE: func(cmd *cobra.Command, args []string) error {
-		return verifyConfigureBridgeParams(configureBridgeParams)
-	},
 	RunE: func(cmd *cobra.Command, args []string) error {
-		endpoint, apiToken, err := credentialmanager.NewCredentialManager(assumeYes).GetCreds(namespace)
-		if err != nil {
-			return errors.New(authErrorMsg)
-		}
-
-		endpoint.Path = path.Join(endpoint.Path, "/v1/config/bridge")
-
-		return configureBridge(endpoint.String(), apiToken, configureBridgeParams)
+		fmt.Println("This command is not supported anymore!")
+		fmt.Println()
+		fmt.Println("You can get your login credentials e.g. by using the following kubectl commands:")
+		fmt.Println("Username - kubectl get secret -n keptn bridge-credentials -o jsonpath=\"{.data.BASIC_AUTH_USERNAME}\" | base64 --decode")
+		fmt.Println("Password - kubectl get secret -n keptn bridge-credentials -o jsonpath=\"{.data.BASIC_AUTH_PASSWORD}\" | base64 --decode")
+		fmt.Println()
+		fmt.Println("For editing the login credentials please use: 'kubectl edit secrets -n keptn bridge-credentials'")
+		fmt.Println()
+		fmt.Println("The URL to your Keptn Bridge can be retrieved using 'keptn status'")
+		return nil
 	},
-}
-
-func configureBridge(endpoint string, apiToken string, configureBridgeParams *configureBridgeCmdParams) error {
-	if configureBridgeParams.Read != nil && *configureBridgeParams.Read {
-		bridgeEndpoint := getBridgeURLFromAPIURL(endpoint)
-		if bridgeEndpoint != "" {
-			fmt.Println("Your Keptn Bridge is available under: " + bridgeEndpoint)
-		}
-		creds, err := retrieveBridgeCredentials(endpoint, apiToken)
-		if err != nil {
-			return err
-		}
-
-		fmt.Println("\nThese are your credentials")
-		fmt.Println("user: " + creds.User)
-		fmt.Println("password: " + creds.Password)
-		return nil
-	}
-	return configureBridgeCredentials(endpoint, apiToken, configureBridgeParams)
-}
-
-func getBridgeURLFromAPIURL(endpoint string) string {
-	parsedURL, err := url.Parse(endpoint)
-	if err != nil {
-		return ""
-	}
-	return strings.Replace(parsedURL.String(), parsedURL.Path, "/bridge", 1)
-}
-
-func retrieveBridgeCredentials(endpoint string, apiToken string) (*configureBridgeAPIPayload, error) {
-	client := &http.Client{
-		Transport: &http.Transport{
-			TLSClientConfig: &tls.Config{InsecureSkipVerify: true}, //nolint:gosec
-		},
-	}
-	req, err := http.NewRequest("GET", endpoint, nil)
-	if err != nil {
-		return nil, err
-	}
-	req.Header.Add("x-token", apiToken)
-	req.Header.Add("content-type", "application/json")
-
-	resp, err := client.Do(req)
-	if err != nil {
-		fmt.Println("Could not send request: " + err.Error())
-		return nil, err
-	}
-	defer resp.Body.Close()
-	body, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		return nil, err
-	}
-
-	if resp.StatusCode < 200 || resp.StatusCode >= 300 {
-		return nil, errors.New("Received not successful response: " + string(body))
-	}
-
-	res := &configureBridgeAPIPayload{}
-	if err := json.Unmarshal(body, res); err != nil {
-		return nil, err
-	}
-	return res, nil
-}
-
-func configureBridgeCredentials(endpoint string, apiToken string, configureBridgeParams *configureBridgeCmdParams) error {
-	bridgeCredentials := configureBridgeAPIPayload{
-		User:     *configureBridgeParams.User,
-		Password: *configureBridgeParams.Password,
-	}
-
-	payload, err := json.Marshal(bridgeCredentials)
-	if err != nil {
-		fmt.Println("Could not marshal response payload: " + err.Error())
-		return err
-	}
-	client := &http.Client{
-		Transport: &http.Transport{
-			TLSClientConfig: &tls.Config{InsecureSkipVerify: true}, //nolint:gosec
-		},
-	}
-
-	req, err := http.NewRequest("POST", endpoint, bytes.NewReader(payload))
-	if err != nil {
-		return err
-	}
-	req.Header.Add("x-token", apiToken)
-	req.Header.Add("content-type", "application/json")
-
-	resp, err := client.Do(req)
-	if err != nil {
-		fmt.Println("Could not send request: " + err.Error())
-		return err
-	}
-	defer resp.Body.Close()
-	body, err := ioutil.ReadAll(resp.Body)
-
-	if resp.StatusCode < 200 || resp.StatusCode >= 300 {
-		return errors.New("Received not successful response: " + string(body))
-	}
-
-	if err != nil {
-		return errors.New("Could not configure bridge: " + err.Error())
-	}
-	fmt.Println("Bridge credentials configured successfully")
-	return nil
-}
-
-func verifyConfigureBridgeParams(configureBridgeParams *configureBridgeCmdParams) error {
-	if configureBridgeParams.Read != nil && *configureBridgeParams.Read {
-		return nil
-	}
-	if configureBridgeParams.User == nil || *configureBridgeParams.User == "" {
-		return errors.New("please specify a user name for exposing the bridge using the '--user=<username>' flag")
-	}
-	if configureBridgeParams.Password == nil || *configureBridgeParams.Password == "" {
-		return errors.New("please specify a password for exposing the bridge using the '--password=<password>' flag")
-	}
-	return nil
 }
 
 func init() {
 	configureCmd.AddCommand(bridgeCmd)
 	configureBridgeParams = &configureBridgeCmdParams{}
-
+	configureBridgeParams.Read = bridgeCmd.Flags().BoolP("output", "o", false, "Print the current credentials")
 	configureBridgeParams.User = bridgeCmd.Flags().StringP("user", "u", "", "The user name to login to the bridge")
 	configureBridgeParams.Password = bridgeCmd.Flags().StringP("password", "p", "", "The password to login to the bridge")
-
-	configureBridgeParams.Read = bridgeCmd.Flags().BoolP("output", "o", false, "Print the current credentials")
 }
